+++ src/get_auth0_v2_users_by_email_b.d.ts
@@ -5,5 +5,7 @@ export declare type get_auth0_v2_users_by_email_params_I = {
 	email:string;
 	AUTH0_DOMAIN?:string;
 };
-export declare type get_auth0_v2_users_by_email_T = (params:get_auth0_v2_users_by_email_params_I)=>Promise<Response>;
+export declare type get_auth0_v2_users_by_email_T = (
+	params:get_auth0_v2_users_by_email_params_I
+)=>Promise<[get_auth0_v2_users_by_email_response_T, Response]>;
 export declare type get_auth0_v2_users_by_email_response_T = Auth0UserProfile[];