+++ src/patch_auth0_v2_user_b.js
@@ -1,5 +1,5 @@
 import { verify_access_token_header_authorization_, AUTH0_DOMAIN$_b } from '@ctx-core/auth0'
-import { safe_fetch as fetch } from '@ctx-core/fetch-undici'
+import { fetch } from '@ctx-core/fetch-undici'
 import { be_ } from '@ctx-core/object'
 import { auth0_management_token__b } from './auth0_management_token__b.js'
 const key = 'patch_auth0_v2_user'
@@ -18,7 +18,7 @@ export const patch_auth0_v2_user_b = be_(key, ctx=>{
 		const auth0_management_token = await auth0_management_token_()
 		const authorization = verify_access_token_header_authorization_(auth0_management_token)
 		const url = `https://${AUTH0_DOMAIN$.$}/api/v2/users/${user_id}`
-		return fetch(url, {
+		const res = await fetch(url, {
 			method: 'PATCH',
 			headers: {
 				'Content-Type': 'application/json',
@@ -26,5 +26,6 @@ export const patch_auth0_v2_user_b = be_(key, ctx=>{
 			},
 			body: JSON.stringify(data)
 		})
+		return [await res.json(), res]
 	}
 })